requirements_kaggle.txt
torch==2.1.0
torchvision==0.16.0
diffusers==0.25.0
accelerate==0.31.0
transformers==4.43.3
onnxruntime-gpu==1.19.2
ruamel.yaml
lpips
torch-fidelity
opencv-python
scikit-image
peft
huggingface_hub
python-dotenv
psutil
matplotlib
Pillow
